requirements.txt
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.39.0
accelerate>=0.28.0
pandas>=2.0.0
numpy>=1.24.0
Pillow>=10.0.0
tqdm>=4.66.0
PyYAML>=6.0.0
evaluate>=0.4.1
nltk>=3.8.1
kagglehub>=0.2.0
sentencepiece>=0.1.99
